backend/gtfs-consumer/requirements.txt
protobuf==4.25.1
requests==2.31.0
kafka-python==2.0.2
gtfs-realtime-bindings==1.0.0
python-dotenv==1.0.0
